[APT-9543] Refactor common pieces for header-aware media
