+++ Armadillo/src/main/java/com/scribd/armadillo/playback/mediasource/HeadersMediaSourceHelper.kt
@@ -1,0 +1,10 @@
+package com.scribd.armadillo.playback.mediasource
+
+import android.content.Context
+import com.google.android.exoplayer2.upstream.DataSource
+import com.scribd.armadillo.models.AudioPlayable
+
+internal interface HeadersMediaSourceHelper {
+    fun createDataSourceFactory(context: Context, request: AudioPlayable.MediaRequest): DataSource.Factory
+    fun updateMediaSourceHeaders(request: AudioPlayable.MediaRequest)
+}
